Allow snpEff errors to be reported when filtering
By default, snpEff does not gzip the output. If we pipe the snpEff
output to gzip, the return code and error messages from snpEff get
overwritten by gzip, and we no longer detect when there's a problem
with snpEff. Instead of piping the snpEff output to gzip, just keep
the default output from snpEff in unzipped form. These temporary
files are deleted in the end anyway. It's more important to know
when snpEff fails.

Note: with snpEff version 4.3, they do now accept input that has
been gzipped. They also have a -o option for the output, but I'm
not sure when that option was introduced, so I'm leaving the
command as is where it writes to stdout and is piped to the
output file.