Update the 'SS' field to be compliant with TCGA VCF version 1.1
The TCGA VCF specification for the Somatic Status or 'SS'
field changed from version 1.0 to version 1.1.

TCGA VCF version 1.0:
Variant status relative to non-adjacent Normal,
0=wildtype,1=germline,2=somatic,3=LOH,4=unknown

TCGA VCF version 1.1:
Variant status relative to non-adjacent Normal,
0=wildtype,1=germline,2=somatic,3=LOH,4=post-transcriptional
modification,5=unknown

Since RADIA was focused on the RNA from the start, it added a
variant status for RNA-Editing as follows:
Variant status relative to non-adjacent Normal,
0=wildtype,1=germline,2=somatic,3=LOH,4=unknown,5=rnaEditing

This fix updates the 'SS' field to be compliant with the TCGA
VCF version 1.1 and handles all of the downstream filtering
that may have used the 'SS' field.