Deleted outdated code that re-queried base quality scores
Around 2 years ago, some RADIA results were generated
without using the BAQ (Base Alignment Quality) scores from
samtools. Then the switch was made to use BAQ which is
recommended by default, and the base quality scores for
the old results needed to be updated. Instead of re-running
the whole pipeline, it was faster to just re-filter and
adjust the base quality scores only for the passing somatic
calls. This code hasn't been used in years and is no longer
needed.