When filtering, check for empty datasets with '.' or './.' genotypes
If a sample has no data, then the default genotype is '.' for
haploid calls (e.g. chrom Y) and './.' for diploid calls. When
filtering, the genotype field is checked to determine if a
particular sample has data or not. We need to check for either
'.' or './.' now instead of just checking for '.'.